+++ src/main/java/dhrlang/lexer/Lexer.java
@@ -73,6 +73,7 @@ public class Lexer {
         keywords.put("@nonreentrant", TokenType.NONREENTRANT);
         keywords.put("@constructor", TokenType.CONSTRUCTOR);
         keywords.put("@event", TokenType.EVENT);
+        keywords.put("@error", TokenType.ERROR);
         keywords.put("emit", TokenType.EMIT);
         keywords.put("@immutable", TokenType.IMMUTABLE);
         keywords.put("@invariant", TokenType.INVARIANT);